Rearchitecture (smaller ROM, non-blocking mode):
Most static information is now in ROM and detected by swimcat.py there.
It has pointers to the respective RAM locations.
This removes the need to keep a copy of the (large) RAM initializer in
ROM.

(Non)Blocking mode is implemented via a flag that swimcat is setting
upon connecting.